refactor mmedit/models/utils/
