+++ mmedit/models/editors/edsr/edsr_net.py
@@ -5,7 +5,8 @@ import torch
 import torch.nn as nn
 from mmengine.model import BaseModule
 
-from mmedit.models.utils import PixelShufflePack, ResidualBlockNoBN, make_layer
+from mmedit.models.layers import PixelShufflePack, ResidualBlockNoBN
+from mmedit.models.utils import make_layer
 from mmedit.registry import MODELS
