+++ mmedit/models/editors/mspie/mspie_stylegan2.py
@@ -15,7 +15,7 @@ from torch import Tensor
 
 from mmedit.registry import MODELS
 from mmedit.structures import EditDataSample
-from ...utils import gather_log_vars, set_requires_grad
+from ...utils import set_requires_grad
 from ..stylegan2 import StyleGAN2
 
 ModelType = Union[Dict, nn.Module]
@@ -98,7 +98,7 @@ class MSPIEStyleGAN2(StyleGAN2):
                         **data, optimizer_wrapper=gen_optimizer_wrapper)
 
                 log_vars_gen_list.append(log_vars_gen)
-            log_vars_gen = gather_log_vars(log_vars_gen_list)
+            log_vars_gen = self.gather_log_vars(log_vars_gen_list)
             log_vars_gen.pop('loss', None)  # remove 'loss' from gen logs
 
             set_requires_grad(self.discriminator, True)
