+++ mmedit/models/editors/stylegan2/stylegan2.py
@@ -14,7 +14,7 @@ from mmedit.registry import MODELS, MODULES
 from mmedit.structures import EditDataSample
 from ...base_models import BaseGAN
 from ...losses import gen_path_regularizer, r1_gradient_penalty_loss
-from ...utils import gather_log_vars, set_requires_grad
+from ...utils import set_requires_grad
 
 ModelType = Union[Dict, nn.Module]
 
@@ -280,7 +280,7 @@ class StyleGAN2(BaseGAN):
                                                     gen_optimizer_wrapper)
 
                 log_vars_gen_list.append(log_vars_gen)
-            log_vars_gen = gather_log_vars(log_vars_gen_list)
+            log_vars_gen = self.gather_log_vars(log_vars_gen_list)
             log_vars_gen.pop('loss', None)  # remove 'loss' from gen logs
 
             set_requires_grad(self.discriminator, True)
