+++ mmedit/models/editors/ttsr/ttsr_net.py
@@ -6,7 +6,8 @@ import torch.nn.functional as F
 from mmcv.cnn import build_conv_layer
 from mmengine.model import BaseModule
 
-from mmedit.models.utils import PixelShufflePack, ResidualBlockNoBN, make_layer
+from mmedit.models.layers import PixelShufflePack, ResidualBlockNoBN
+from mmedit.models.utils import make_layer
 from mmedit.registry import MODELS
 
 # Use partial to specify some default arguments
