+++ mmedit/models/layers/__init__.py
@@ -1,13 +1,23 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 # To register Deconv
+from .all_gather_layer import AllGatherLayer
 from .aspp import ASPP
 from .conv import *  # noqa: F401, F403
 from .conv2d_gradfix import conv2d, conv_transpose2d
+from .downsample import pixel_unshuffle
+from .ensemble import SpatialTemporalEnsemble
 from .gated_conv_module import SimpleGatedConvModule
+from .img_normalize import ImgNormalize
 from .linear_module import LinearModule
 from .separable_conv_module import DepthwiseSeparableConvModule
+from .sr_backbone import ResidualBlockNoBN
+from .upsample import PixelShufflePack
+
+# from .sr_backbone import ResidualBlockNoBN, default_init_weights
 
 __all__ = [
     'ASPP', 'DepthwiseSeparableConvModule', 'SimpleGatedConvModule',
-    'LinearModule', 'conv2d', 'conv_transpose2d'
+    'LinearModule', 'conv2d', 'conv_transpose2d', 'pixel_unshuffle',
+    'PixelShufflePack', 'ImgNormalize', 'SpatialTemporalEnsemble',
+    'AllGatherLayer', 'ResidualBlockNoBN'
 ]
