+++ mmedit/models/layers/sr_backbone.py
@@ -1,42 +1,40 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 import torch.nn as nn
-from mmengine.model.weight_init import constant_init, kaiming_init
-from mmengine.utils.dl_utils.parrots_wrapper import _BatchNorm
 
+from ..utils import default_init_weights
 
-def default_init_weights(module, scale=1):
-    """Initialize network weights.
+# def default_init_weights(module, scale=1):
+#     """Initialize network weights.
 
-    Args:
-        modules (nn.Module): Modules to be initialized.
-        scale (float): Scale initialized weights, especially for residual
-            blocks. Default: 1.
-    """
-    for m in module.modules():
-        if isinstance(m, nn.Conv2d):
-            kaiming_init(m, a=0, mode='fan_in', bias=0)
-            m.weight.data *= scale
-        elif isinstance(m, nn.Linear):
-            kaiming_init(m, a=0, mode='fan_in', bias=0)
-            m.weight.data *= scale
-        elif isinstance(m, _BatchNorm):
-            constant_init(m.weight, val=1, bias=0)
+#     Args:
+#         modules (nn.Module): Modules to be initialized.
+#         scale (float): Scale initialized weights, especially for residual
+#             blocks. Default: 1.
+#     """
+#     for m in module.modules():
+#         if isinstance(m, nn.Conv2d):
+#             kaiming_init(m, a=0, mode='fan_in', bias=0)
+#             m.weight.data *= scale
+#         elif isinstance(m, nn.Linear):
+#             kaiming_init(m, a=0, mode='fan_in', bias=0)
+#             m.weight.data *= scale
+#         elif isinstance(m, _BatchNorm):
+#             constant_init(m.weight, val=1, bias=0)
 
+# def make_layer(block, num_blocks, **kwarg):
+#     """Make layers by stacking the same blocks.
 
-def make_layer(block, num_blocks, **kwarg):
-    """Make layers by stacking the same blocks.
+#     Args:
+#         block (nn.module): nn.module class for basic block.
+#         num_blocks (int): number of blocks.
 
-    Args:
-        block (nn.module): nn.module class for basic block.
-        num_blocks (int): number of blocks.
-
-    Returns:
-        nn.Sequential: Stacked blocks in nn.Sequential.
-    """
-    layers = []
-    for _ in range(num_blocks):
-        layers.append(block(**kwarg))
-    return nn.Sequential(*layers)
+#     Returns:
+#         nn.Sequential: Stacked blocks in nn.Sequential.
+#     """
+#     layers = []
+#     for _ in range(num_blocks):
+#         layers.append(block(**kwarg))
+#     return nn.Sequential(*layers)
 
 
 class ResidualBlockNoBN(nn.Module):
