+++ mmedit/models/utils/__init__.py
@@ -1,44 +1,24 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .bbox_utils import extract_around_bbox, extract_bbox_patch
-from .dist_utils import AllGatherLayer
-from .downsample import pixel_unshuffle
-from .ensemble import SpatialTemporalEnsemble
 from .flow_warp import flow_warp
-from .generation_model_utils import (GANImageBuffer, ResidualBlockWithDropout,
-                                     UnetSkipConnectionBlock,
-                                     generation_init_weights,
-                                     get_module_device, get_valid_noise_size,
-                                     get_valid_num_batches, set_requires_grad)
-from .img_normalize import ImgNormalize
-from .log_utils import gather_log_vars
+from .model_utils import (default_init_weights, generation_init_weights,
+                          get_module_device, get_valid_noise_size,
+                          get_valid_num_batches, make_layer, set_requires_grad)
 from .sampling_utils import label_sample_fn, noise_sample_fn
-from .sr_backbone_utils import (ResidualBlockNoBN, default_init_weights,
-                                make_layer)
 from .tensor_utils import get_unknown_tensor
-from .upsample import PixelShufflePack
 
 __all__ = [
-    'PixelShufflePack',
     'default_init_weights',
-    'ResidualBlockNoBN',
     'make_layer',
     'flow_warp',
-    'ImgNormalize',
     'generation_init_weights',
-    'GANImageBuffer',
-    'UnetSkipConnectionBlock',
-    'ResidualBlockWithDropout',
-    'pixel_unshuffle',
-    'SpatialTemporalEnsemble',
     'set_requires_grad',
     'extract_bbox_patch',
     'extract_around_bbox',
     'get_unknown_tensor',
-    'gather_log_vars',
     'noise_sample_fn',
     'label_sample_fn',
     'get_valid_num_batches',
     'get_valid_noise_size',
     'get_module_device',
-    'AllGatherLayer',
 ]
