--- tests/test_models/test_utils/test_dist_utils.py
@@ -1,1 +1,0 @@
-# Copyright (c) OpenMMLab. All rights reserved.
